requirements.txt
langchain
langchain-community
youtube-transcript-api
faiss-cpu
tiktoken
openai
python-dotenv
